lots of updates
made a function call for displayedSavedRecipes. Also added an id to the container of displayResults so that it can be targeted correctly. Also fixed some bugs with the saveItemToSavedRecipes function that were appearing due to new elements being added that made the bubbling error. Moved contents of script tag at bottom of savedRecipes into script js. These are outside of the jquery function call so that they can use html onclick. Also added some margin to the bottom of results that will be displayed on results page. Also removed a redundant call to jquery. also added some margin to generated saved recipes.